requirements.txt
pdf2image==1.16.0
Pillow==9.3.0
PyPDF2==3.0.1
